webapp/pyproject.toml
[tool.poetry]
name = "crowdeval"
version = "0.1.0"
description = ""
authors = ["Freddy Heppell <freddy@freddyheppell.com>"]

[tool.poetry.dependencies]
python = "3.9.4"
Flask = "^1.1.2"
environs = "^9.3.1"
Flask-Static-Digest = "^0.2.1"
Flask-Dance = {extras = ["sqla"], version = "^3.2.0"}
Flask-SQLAlchemy = "^2.4.4"
Flask-Login = "^0.5.0"
Flask-Migrate = "^2.6.0"
blinker = "^1.4"
Flask-DebugToolbar = "^0.11.0"
mysql-connector-python = "^8.0.23"
Flask-WTF = "^0.14.3"
TwitterAPI = "^2.6.9"
Flask-BS4 = "^5.0.0"
Flask-Elasticsearch = "^0.2.5"
bert-serving-client = "^1.10.0"
timeago = "^1.0.15"
Flask-Caching = "^1.10.1"
redis = "^3.5.3"
Faker = "^6.5.2"
factory-boy = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pylint = "^2.6.0"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-blind-except = "^0.2.0"
flake8-debugger = "^4.0.0"
flake8-docstrings = "^1.5.0"
flake8-isort = "^4.0.0"
isort = "^5.7.0"
pep8-naming = "^0.11.1"
WebTest = "^2.0.35"
pytest-mock = "^3.5.1"
mock = "^4.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip_glob = ["migrations/*", "seeds/*"]

[pytest]
mock_use_standalone_module = true